+++ src/widget/compact/compactpicker.js
@@ -2,12 +2,14 @@ import Widget from '../../js/widget';
 import { getSiblingElements } from '../../js/dom-utils';
 
 /**
- * Compact Picker. Hides text labels if a media label is present.
+ * Media Picker. Hides text labels if a media label is present.
+ * 
+ * TODO: change the name of this widget
  */
 class CompactPicker extends Widget {
 
     static get selector() {
-        return '.or-appearance-compact, .or-appearance-quickcompact, [class*="or-appearance-compact-"]';
+        return '.or-appearance-no-buttons';
     }
 
     _init() {