+++ src/widget/horizontal-choices/horizontalchoices.js
@@ -3,11 +3,13 @@ import Widget from '../../js/widget';
 /**
  * Horizontal Choices Widgets. Adds a filler if the last row contains two elements.
  * The filler avoids the last radiobutton or checkbox to not be lined up correctly below the second column.
+ * 
+ * Change the name of this widget to Columns or something
  */
 class HorizontalChoices extends Widget {
 
     static get selector() {
-        return '.question.or-appearance-horizontal';
+        return '.question.or-appearance-columns';
     }
 
     _init() {
@@ -21,7 +23,7 @@ class HorizontalChoices extends Widget {
                 fillers--;
             }
             // if added to correct question type, add initialized class
-            this.question.classList.add( 'or-horizontal-initialized' );
+            this.question.classList.add( 'or-columns-initialized' );
         } );
     }
 }