+++ CHANGELOG.md
@@ -5,6 +5,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 [Unreleased]
 --------------------
+##### Removed
+- Support for appearances "quick" and "quickcompact".
+
+##### Added
+- Support for appearances "columns", "columns-flex", "columns-n" all with an optional "no-buttons" modififier
+
 ##### Fixed
 - Automatically enlarged multiline inputs cannot be manually resized to their original size.